courses page almost done, coursedetails page basic components done
